backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.2
django-cors-headers>=4.4
python-dotenv>=1.0
pytest>=8.0
pytest-django>=4.8
reportlab>=4.0
